Fixed two problems with annotated types in patterns
Problem 1: The parser did not accept them. It has to accept a "RefinedType" as an ascription,
not a "WithType" (as it did before), or even a "SimpleType" (as speced in the SyntaxSummary).

Problem 2: Annotations are always typed as expressions. The annotations in question were typed
as patterns before.

Tests in Patterns.scala and in the Dotty compiler itself.